refactor: move error response handling to responseHandler
