+++ packages/temba/src/index.ts
@@ -1,7 +1,8 @@
 import { createServer as httpCreateServer } from 'node:http'
 import { initConfig, type UserConfig } from './config'
 import type { IncomingMessage, ServerResponse } from 'http'
-import { createResourceHandler, handleNotFound, sendErrorResponse } from './resourceHandler'
+import { createResourceHandler } from './resourceHandler'
+import { handleNotFound, sendErrorResponse } from './responseHandler'
 import { getHttpLogger, initLogger } from './log/logger'
 import { createOpenApiHandler, getOpenApiPaths } from './openapi'
 import { TembaError as TembaErrorInternal } from './requestInterceptor/TembaError'
